refactor: waiting 패키지 이동
