+++ src/main/java/roomescape/reservation/controller/response/MyReservationResponse.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.domain.ReservationWaiting;
+import roomescape.waiting.domain.ReservationWaiting;
 import roomescape.waiting.dto.WaitingWithRank;
 
 public record MyReservationResponse(Long id,
